+++ NEWS.md
@@ -2,3 +2,4 @@
 
 * Added a `set_source_script()`
 * You can now also supply an email address to `dsi_sources_from_paths()`
+* add `sources()` accessors